+++ Services/skill.py
@@ -1,7 +1,6 @@
 from sqlalchemy.orm import Session
 from fastapi import HTTPException
 from Repositories.skill import SkillRepository
-from Repositories.user import UserRepository
 from Schemas.skill import SkillCreate, SkillUpdate
 from uuid import UUID
 
@@ -9,10 +8,8 @@ class SkillService:
     def __init__(self, db: Session):
         self.repository = SkillRepository(db)
 
-    def create_skill(self, skill_data: SkillCreate):
-        user_repo = UserRepository(self.repository.db)
-        if not user_repo.get_by_id(skill_data.user_id):
-            raise HTTPException(status_code=400, detail="Usuário associado não existe")
+    def create_skill(self, skill_data: SkillCreate, current_user_id: UUID):
+        skill_data.user_id = current_user_id
         return self.repository.create(skill_data)
 
     def get_all_skills(self):
@@ -25,19 +22,20 @@ class SkillService:
         return skill
 
     def get_skills_by_user_id(self, user_id: UUID):
-        user_repo = UserRepository(self.repository.db)
-        if not user_repo.get_by_id(user_id):
-            raise HTTPException(status_code=404, detail="Usuário não encontrado")
         return self.repository.get_by_user_id(user_id)
 
-    def update_skill(self, skill_id: UUID, skill_data: SkillUpdate):
+    def update_skill(self, skill_id: UUID, skill_data: SkillUpdate, current_user_id: UUID):
+        skill = self.get_skill_by_id(skill_id)
+        if skill.user_id != current_user_id:
+            raise HTTPException(status_code=403, detail="Sem permissão para editar essa skill")
+
         updated = self.repository.update(skill_id, skill_data)
-        if not updated:
-            raise HTTPException(status_code=404, detail="Skill não encontrada")
         return updated
 
-    def delete_skill(self, skill_id: UUID):
-        success = self.repository.delete(skill_id)
-        if not success:
-            raise HTTPException(status_code=404, detail="Skill não encontrada")
+    def delete_skill(self, skill_id: UUID, current_user_id: UUID):
+        skill = self.get_skill_by_id(skill_id)
+        if skill.user_id != current_user_id:
+            raise HTTPException(status_code=403, detail="Sem permissão para excluir essa skill")
+
+        self.repository.delete(skill_id)
         return {"detail": "Skill deletada com sucesso"}
